adding components for each phase
